requirements.txt
ConfigArgParse==1.5.1
python-dotenv==0.18.0
aiofiles>=0.7.0